A new option "auto" for "knots".
"knots" can be determined automatically when "knots="auto"". The major
work is to add two wrapper functions (cv.kernel.spline.wrapper,
cv.factor.spline.wrapper) to deal with the new option of "knots". For
the exhaustive search of the kernel method (krscv), we will have one
more time to call the objective function to know the type of "knots",
the reason is that "optim" cannot return the attributes.