Add a class to redirect C++ output to Rprintf, add a method to the class Parameters.
std::cout and std::cerr have been redirect to rout which calls Rprintf
to output. We also added a method which is another definition of
Parameters::read except the input parameter which is iostream. Now
snomadr will not write options into a file. Options will be written
into an object of iostream, then NOMAD will read options from the
object of iostream.